pyproject.toml
[tool.poetry]
name = "sb-surfshark-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
python-json-logger = "^2.0.7"
bs4 = "^0.0.2"
requests = "^2.31.0"
pandas = "^2.2.2"
lxml = "^5.2.1"
openpyxl = "^3.1.2"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
fastparquet = "^2024.2.0"
retrying = "^1.3.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
